requirements.txt
boto
simplejson>=2.1
python-dateutil>=2.1
unittest2>=0.5.1
# -e git+https://github.com/martinpaljak/M2Crypto.git#egg=M2Crypto
M2Crypto>=0.22.3
